feat: moved row component
