+++ src/components/Row/Row.tsx
@@ -1,13 +1,13 @@
 import React, { FC } from 'react';
-import * as Styled from '../EmployeesProfilePage.styles';
+import * as Styled from './Row.styles';
 import { RowProps } from './Row.type';
 
 export const Row: FC<RowProps> = ({ children, title }) => {
   return (
     <>
       <Styled.RowWrapper>
-        <Styled.RowTitleTypography>{title}</Styled.RowTitleTypography>
-        <Styled.RowContentTypography>{children}</Styled.RowContentTypography>
+        <Styled.RowTitle>{title}</Styled.RowTitle>
+        <Styled.RowContent>{children}</Styled.RowContent>
       </Styled.RowWrapper>
     </>
   );
