--- src/components/pages/ProjectsDetailsPage/Row/Row.tsx
@@ -1,14 +1,0 @@
-import React, { FC } from 'react';
-import * as Styled from '../../EmployeesProfilePage/EmployeesProfilePage.styles';
-import { RowProps } from './Row.type';
-
-export const Row: FC<RowProps> = ({ children, title }) => {
-  return (
-    <>
-      <Styled.RowWrapper>
-        <Styled.RowTitleTypography>{title}</Styled.RowTitleTypography>
-        <Styled.RowContentTypography>{children}</Styled.RowContentTypography>
-      </Styled.RowWrapper>
-    </>
-  );
-};
