--- src/components/pages/ProjectsDetailsPage/Row/Row.type.ts
@@ -1,6 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface RowProps {
-  children: ReactNode;
-  title: string;
-}
